Minor changes to avoid some warnings on GCC 4.7.1 (Debian wheezy/sid).  ANDY - please look at the diff and make sure I didn't do something stupid

+++ src/theory/uf/theory_uf_strong_solver.cpp
@@ -463,7 +463,7 @@ void StrongSolverTheoryUf::ConflictFind::getDisequalitiesToRegions( int ri, std:
 
 void StrongSolverTheoryUf::ConflictFind::explainClique( std::vector< Node >& clique, OutputChannel* out ){
   Assert( d_cardinality>0 );
-  while( clique.size()>long(d_cardinality+1) ){
+  while( clique.size()>size_t(d_cardinality+1) ){
     clique.pop_back();
   }
   //found a clique
@@ -490,7 +490,7 @@ void StrongSolverTheoryUf::ConflictFind::explainClique( std::vector< Node >& cli
       conflict.push_back( d_disequalities[i] );
       nodesWithinRep[r1][ d_disequalities[i][0][0] ] = true;
       nodesWithinRep[r2][ d_disequalities[i][0][1] ] = true;
-      if( conflict.size()==((int)clique.size()*( (int)clique.size()-1 )/2) ){
+      if( conflict.size()==(clique.size()*( clique.size()-1 )/2) ){
         break;
       }
     }
